+++ opencog/cython/opencog/forwardchainer.pyx
@@ -1,7 +1,7 @@
 from opencog.atomspace import types
 from cython.operator cimport dereference as deref, preincrement as inc
 from opencog.atomspace cimport cHandle, Atom, AtomSpace, TruthValue
-from forwardchainer cimport cForwardChainer
+from ure cimport cForwardChainer
 
 # Create a Cython extension type which holds a C++ instance
 # as an attribute and create a bunch of forwarding methods
