+++ tests/cython/backwardchainer/test_bc.py
@@ -3,7 +3,7 @@ import unittest
 from unittest import TestCase
 from opencog.scheme_wrapper import scheme_eval
 from opencog.atomspace import TruthValue
-from opencog.backwardchainer import BackwardChainer
+from opencog.ure import BackwardChainer
 from opencog.type_constructors import *
 from opencog.utilities import initialize_opencog
 import __main__
